Iniciando uma classe para testar itens matematicos

+++ Basic/src/xyz/infodata/Math.java
@@ -1,20 +1,23 @@
 package xyz.infodata;
-
+/*
+ * Autor: Nicolas Ramos
+ * Nota : Iniciando uma classe para incluir itens matematicos.
+ */
 public class Math {
 
     // recursivo
-    public long fatorial(long n){
-        if(n < 0L) { return -1L;}
-        if(n == 0L || n == 1L) { return 1L;}
-        return n * fatorial(n-1L);
+    public double fatorial(double n){
+        if(n < 0.0) { return -1.0;}
+        if(n == 0.0 || n == 1.0) { return 1.0;}
+        return n * fatorial(n-1.0);
     }
 
     // interativo
-    public long factorial(long n) {
-        if(n < 0L) { return -1L;}
-        if(n == 0L || n == 1L) { return 1L;}
-        long factorial = 1;
-        for(long i=2; i<=n; i++) {
+    public double factorial(double n) {
+        if(n < 0.0) { return -1.0;}
+        if(n == 0.0 || n == 1.0) { return 1.0;}
+        double factorial = 1;
+        for(double i=2; i<=n; i++) {
             factorial *= i;
         }
         return factorial;
